Move area functions to geometry file
