--- acircle.js
@@ -1,5 +1,0 @@
-var radioBase = parseInt(process.argv[2])
-console.log("the radio of the base is " + radioBase)
-
-var area = (radioBase * radioBase) * 3.14
-console.log("the volume is " + volume)
