--- arect.js
@@ -1,7 +1,0 @@
-var base = parseInt(process.argv[2])
-var height = parseInt(process.argv[3])
-console.log(`the base's length is: ${base}`)
-console.log(`the height's length is: ${height}`)
-
-var area = base * height
-console.log(`the area is: ${area}`)
