--- asquare.js
@@ -1,5 +1,0 @@
-var side = parseInt(process.argv[2])
-console.log(`the side's length is: ${side}`)
-
-var area = side * side
-console.log(`the area is: ${area}`)
